fastapi-backend/requirements.txt

fastapi==0.110.0
uvicorn==0.30.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.6.4
pytest==7.4.3
httpx==0.25.2
